Move generic methods to k8s_util and some cosmetic changes
