+++ istioctl/pkg/install/verifier.go
@@ -35,7 +35,6 @@ import (
 	"istio.io/istio/operator/pkg/controlplane"
 	"istio.io/istio/operator/pkg/translate"
 	"istio.io/istio/operator/pkg/util"
-	"istio.io/istio/pkg/config/schema"
 	"istio.io/pkg/log"
 )
 
@@ -99,7 +98,7 @@ func (v *StatusBasedVerifier) verifyInstall() error {
 }
 
 // nolint: lll
-func (v *StatusBasedVerifier) verifyPostInstallIstioOperator( /*istioNamespaceFlag string, iop *v1alpha1.IstioOperator, filename string, restClientGetter genericclioptions.RESTClientGetter*/ iop *v1alpha1.IstioOperator, filename string) (int, int, error) {
+func (v *StatusBasedVerifier) verifyPostInstallIstioOperator(iop *v1alpha1.IstioOperator, filename string) (int, int, error) {
 	// Generate the manifest this IstioOperator will make
 	t := translate.NewTranslator()
 
@@ -176,8 +175,7 @@ func (v *StatusBasedVerifier) verifyPostInstall(visitor resource.Visitor, filena
 			if err != nil {
 				return err
 			}
-			err = getDeploymentStatus(deployment)
-			if err != nil {
+			if err = verifyDeploymentStatus(deployment); err != nil {
 				return istioVerificationFailureError(filename, err)
 			}
 			if namespace == v.IstioNamespace && strings.HasPrefix(name, "istio") {
@@ -207,8 +205,7 @@ func (v *StatusBasedVerifier) verifyPostInstall(visitor resource.Visitor, filena
 			// IstioOperator isn't part of pkg/config/schema/collections,
 			// usual conversion not available.  Convert unstructured to string
 			// and ask operator code to unmarshal.
-
-			un.SetCreationTimestamp(meta_v1.Time{}) // UnmarshalIstioOperator chokes on these
+			fixTimestampRelatedUnmarshalIssues(un)
 			by := util.ToYAML(un)
 			iop, err := operator_istio.UnmarshalIstioOperator(by, true)
 			if err != nil {
@@ -243,13 +240,10 @@ func (v *StatusBasedVerifier) verifyPostInstall(visitor resource.Visitor, filena
 				crdCount++
 			}
 		}
-		scope.Debugf("%s: %s.%s checked successfully", kind, name, namespace)
+		scope.Infof("%s: %s.%s checked successfully", kind, name, namespace)
 		return nil
 	})
-	if err != nil {
-		return crdCount, istioDeploymentCount, err
-	}
-	return crdCount, istioDeploymentCount, nil
+	return crdCount, istioDeploymentCount, err
 }
 
 func (v *StatusBasedVerifier) verifyJobPostInstall(job *v1batch.Job) error {
@@ -328,45 +322,6 @@ func reportStatus(crdCount, istioDeploymentCount int, err error) error {
 	}
 	scope.Infof("Istio is installed successfully")
 	return nil
-}
-
-func getDeploymentStatus(deployment *appsv1.Deployment) error {
-	cond := getDeploymentCondition(deployment.Status, appsv1.DeploymentProgressing)
-	if cond != nil && cond.Reason == "ProgressDeadlineExceeded" {
-		return fmt.Errorf("deployment %q exceeded its progress deadline", deployment.Name)
-	}
-	if deployment.Spec.Replicas != nil && deployment.Status.UpdatedReplicas < *deployment.Spec.Replicas {
-		return fmt.Errorf("waiting for deployment %q rollout to finish: %d out of %d new replicas have been updated",
-			deployment.Name, deployment.Status.UpdatedReplicas, *deployment.Spec.Replicas)
-	}
-	if deployment.Status.Replicas > deployment.Status.UpdatedReplicas {
-		return fmt.Errorf("waiting for deployment %q rollout to finish: %d old replicas are pending termination",
-			deployment.Name, deployment.Status.Replicas-deployment.Status.UpdatedReplicas)
-	}
-	if deployment.Status.AvailableReplicas < deployment.Status.UpdatedReplicas {
-		return fmt.Errorf("waiting for deployment %q rollout to finish: %d of %d updated replicas are available",
-			deployment.Name, deployment.Status.AvailableReplicas, deployment.Status.UpdatedReplicas)
-	}
-	return nil
-}
-
-func getDeploymentCondition(status appsv1.DeploymentStatus, condType appsv1.DeploymentConditionType) *appsv1.DeploymentCondition {
-	for i := range status.Conditions {
-		c := status.Conditions[i]
-		if c.Type == condType {
-			return &c
-		}
-	}
-	return nil
-}
-
-func findResourceInSpec(kind string) string {
-	for _, c := range schema.MustGet().KubeCollections().All() {
-		if c.Resource().Kind() == kind {
-			return c.Resource().Plural()
-		}
-	}
-	return ""
 }
 
 func istioVerificationFailureError(filename string, reason error) error {
